requirements.txt
python-telegram-bot==20.7
python-docx==1.1.0
python-dotenv==1.0.1
requests==2.31.0
pillow==10.2.0